Exchanged $ with jQuery to avoid conflicts
In some cases where different JS-libraries are loaded too the slider crashes with the console error message:
"Uncaught TypeError: $(...).load is not a function"
This may be caused by a conflict with the $ from prototype.js. I had this case on a page using tt_news category menu which adds prototype.js after the flexslider's jquery - call and leads to this error.
Also I changed the jQuery.load() - function call with the jQuery.ready() to be sure the DOM is loaded and the accessed selector is available.